Mark points save in Exif; ctrl+wheel zoom
Areas of interest marked with the left-click are now save immediately to
the images exif metadata, under the UserComment section. To avoid
overwriting or otherwise messing up the exif data, I have arbitrarily
set the max number of markers to 5. If you go over five, it will delete
the first one put down (FIFO). Right clicking on the image will remove
all markers from the image. These marker feature required several new
functions in the Tag_Handler class.

Holding the control key and then scrolling the mouse on the image will
zoom the view to the mouse cursor.

Forced reload of the selected image if a list view is clicked. This was
preventing an image from loading when moving back and forth between
priority and available images, but keeping the same images selected.

A little bit of spring cleaning was done here and there as well.